server_flask/requirements.txt
imutils
matplotlib
numpy
Pillow
flask
flask-socketio
gevent
gevent-websocket
pytesseract
tensorflow
